Move AuthConfiguration to `com.amplifyframework.auth.cognito` package
