+++ aws-auth-cognito/src/main/java/com/amplifyframework/statemachine/codegen/events/AuthEvent.kt
@@ -15,9 +15,9 @@
 
 package com.amplifyframework.statemachine.codegen.events
 
+import com.amplifyframework.auth.cognito.AuthConfiguration
 import com.amplifyframework.statemachine.StateMachineEvent
 import com.amplifyframework.statemachine.codegen.data.AmplifyCredential
-import com.amplifyframework.statemachine.codegen.data.AuthConfiguration
 import java.util.Date
 
 internal class AuthEvent(val eventType: EventType, override val time: Date? = null) :
